+++ pkg/apis/core/v1alpha1/config_types_test.go
@@ -1,0 +1,9 @@
+package v1alpha1
+
+import "testing"
+
+func Test_Config_Spec_Status_App(t *testing.T) {
+	t.Log("ensure ConfigSpecApp ConfigStatusApp have identical fields")
+	specApp := ConfigSpecApp{}
+	_ = ConfigStatusApp(specApp)
+}